Split file to ensure correct authorship and license is applied.
